rename Algae pickup to Algae intake
