+++ src/main/java/frc/robot/RobotContainer.java
@@ -32,8 +32,8 @@ import frc.robot.commands.PickupCoral;
 import frc.robot.commands.ReverseAlgaePickup;
 import frc.robot.commands.SetElevatorHeight;
 import frc.robot.commands.SubsystemControl;
+import frc.robot.subsystems.algaeIntake.AlgaeIntake;
 import frc.robot.subsystems.algaeknocker.AlgaeKnocker;
-import frc.robot.subsystems.algaepickup.AlgaeIntake;
 import frc.robot.subsystems.coralScorer.CoralScorer;
 import frc.robot.subsystems.drive.Drive;
 import frc.robot.subsystems.drive.GyroIO;
