+++ src/main/java/frc/robot/commands/AlgaeIntakeExtendPistons.java
@@ -1,7 +1,7 @@
 package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.algaepickup.AlgaeIntake;
+import frc.robot.subsystems.algaeIntake.AlgaeIntake;
 
 public class AlgaeIntakeExtendPistons extends Command {
     private final AlgaeIntake algaeSolenoidRight;
