+++ src/main/java/frc/robot/commands/ReverseAlgaePickup.java
@@ -2,7 +2,7 @@ package frc.robot.commands;
 
 import edu.wpi.first.wpilibj.Timer;
 import edu.wpi.first.wpilibj2.command.Command;
-import frc.robot.subsystems.algaepickup.AlgaeIntake;
+import frc.robot.subsystems.algaeIntake.AlgaeIntake;
 
 public class ReverseAlgaePickup extends Command {
     private final AlgaeIntake algaeIntakeMotor;
