+++ src/main/java/frc/robot/subsystems/algaeIntake/AlgaeIntakeIO.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.algaepickup;
+package frc.robot.subsystems.algaeIntake;
 
 import org.littletonrobotics.junction.AutoLog;
